Production readiness: fix 100+ issues across backend and frontend
Backend (44 files):
- Implement 26 empty router stubs (community, gamification, referrals, reports, analytics, external projects)
- Fix security: OAuth CSRF protection, admin cascade delete, payment auth checks, dispute/gig/video call auth
- Fix payment security: remove direct balance manipulation, add gateway verification
- Fix schemas: field mismatches, missing fields, Float→Numeric(12,2) for monetary values
- Fix InvoiceTaxService: replace all mock data with real DB queries
- Fix TimezoneService: real timezone conversion with zoneinfo
- Fix seed_db.py and migrations for Turso HTTP API
- Fix phone verification SMS integration (Twilio/SNS)
- Fix document upload to actually save files to disk
- Fix Stripe payment methods customer ID lookup
- Fix API key hashing (SHA-256)
- Fix contract_amount sync with amount field

Frontend (39 files):
- Replace ALL mock/hardcoded data with real API calls (Community, Help, Gamification, Referrals, Leaderboard)
- Remove fake setTimeout delays from all 8 AI tools
- Fix localStorage privilege escalation in portal layout
- Fix broken /portal/ URLs in Favorites and Disputes
- Fix hardcoded userId in payout-methods, disputes/create, invoices/create
- Fix admin dashboard mock data (geo distribution, security score)
- Fix Login branding panel to show correct role
- Fix logout handler to call backend and clear all tokens
- Fix duplicate Ctrl+K handlers
- Wire up non-functional buttons in Workroom
- Add error feedback in proposal accept/reject
- Remove console.log from production code
- Fix empty catch blocks with proper error logging
- Replace mock fallbacks in API layer (SimilarJobs, AdminPolicyEditor, ProjectAICopilot)

+++ backend/app/models/dispute.py
@@ -1,6 +1,6 @@
 # @AI-HINT: Dispute model for contract disagreements with admin arbitration workflow
 """Dispute models for MegiLance platform"""
-from sqlalchemy import String, Text, DateTime, ForeignKey, Float
+from sqlalchemy import String, Text, DateTime, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
@@ -48,7 +48,7 @@ class Dispute(Base):
     created_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc), index=True)
     resolved_at: Mapped[datetime] = mapped_column(DateTime, nullable=True)
     resolution: Mapped[str] = mapped_column(Text, nullable=True)
-    resolution_amount: Mapped[float] = mapped_column(Float, nullable=True)
+    resolution_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=True)
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
 
     # Relationships
