+++ backend/app/models/refund.py
@@ -2,7 +2,7 @@
 """
 Refund model for payment reversals
 """
-from sqlalchemy import String, Float, DateTime, ForeignKey
+from sqlalchemy import String, Numeric, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
@@ -21,7 +21,7 @@ class Refund(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     payment_id: Mapped[int] = mapped_column(ForeignKey("payments.id"), index=True)
-    amount: Mapped[float] = mapped_column(Float)
+    amount: Mapped[float] = mapped_column(Numeric(12, 2))
     reason: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
     status: Mapped[str] = mapped_column(String(20), default="pending", index=True)  # pending, approved, rejected, processed
     requested_by: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
